backend/requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
pyjwt
bcrypt==4.0.1
python-multipart
email-validator
slowapi
psycopg2-binary
fastapi-socketio
sqlalchemy
python-socketio
sentencepiece==0.1.99 
transformers==4.41.1
evaluate 
rouge_score 
sympy
typing-extensions==4.12.2
pyarrow
torch_geometric
omegaconf